requirements.txt
biopython
pandas
numpy
openpyxl
tqdm

